Add map date filter to search filter menu
