+++ TODO.md
@@ -200,7 +200,7 @@
 * [BUG] als je /profile window klein maakt kan je niet scrollen naar onder
 * in search filters, gebruik zelfde date range filter UI component als in map view, maar dan fixed, niet openen met
   button
-* Theme picker in profile v-menu top right (same as ruurd music). Dark | Light | Follow system | Follow schedule
+* ✅ Theme picker in profile v-menu top right (same as ruurd music). Dark | Light | Follow system | Follow schedule
 * [investigate] krijg je ratio/monthItem desync als je de map eerst laad, terwijl de backend foto metadata ingest, en
   dan naar de timeline gaat? Want de map date filter heeft dan al de ratios opgevraagd in de timelineStore.
 * settings page heeft geen visible scrollbar
